+++ package.json
@@ -42,7 +42,6 @@
     "react-linkify": "^1.0.0-alpha",
     "react-redux": "^8.1.0",
     "react-router-dom": "^6.12.1",
-    "react-slack-feedback": "^2.1.1",
     "react-tagcloud": "^2.3.1",
     "redux-persist": "^6.0.0",
     "redux-saga": "^1.2.3",
